examples/knowledgebase/requirements.txt
python-dotenv>=1.0.0
neo4j>=5.0.0
asyncio>=3.4.3
pydantic>=2.0.0
